+++ Banking_App_v0.1/Forms/AccountsReceiveMoney.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Banking_App_v0._1
+{
+    public partial class AccountsReceiveMoney : Form
+    {
+        public AccountsReceiveMoney()
+        {
+            InitializeComponent();
+        }
+
+        private void btnConfirmReceive_Click(object sender, EventArgs e)
+        {
+
+            this.Close();
+        }
+
+        private void homePageToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+                this.Hide();
+                Landing landing = new Landing();
+                landing.ShowDialog();
+        }
+    }
+}
