+++ Banking_App_v0.1/Forms/Landing.cs
@@ -19,7 +19,9 @@ namespace Banking_App_v0._1
 
         private void btnLogout_Click(object sender, EventArgs e)
         {
-            this.Close();
+            this.Hide();
+            LoginPage loginPage = new LoginPage();
+            loginPage.ShowDialog();
         }
 
         private void btnAccounts_Click(object sender, EventArgs e)
